PR S4.5: decompose blackjack_cog into _state + _persistence + views/
Largest single decomposition in Phase A.  Cog drops 1702 → 650 LOC
(62% reduction, well under S4.6's 800 ceiling).  Pattern matches the
established F-3 + helpers-package convention used by economy,
moderation, diagnostic, and xp.

New layout:

  cogs/blackjack_cog.py             — BlackjackCog (4 commands +
                                      cog_load + recovery +
                                      on_guild_remove +
                                      on_raw_reaction_add +
                                      _launch_tournament) +
                                      back-compat re-exports for tests
  cogs/blackjack/__init__.py        — package docstring
  cogs/blackjack/_state.py          — data classes (_Game, _PvPState,
                                      _BjTournament, _TournPlayerState),
                                      module dicts (_active, _pvp,
                                      _tournaments), public constants
                                      (FREE_WIN_COINS, TOURN_*,
                                      BLACKJACK_*_SUBSYSTEM/VERSION)
  cogs/blackjack/_persistence.py    — 10 persistence helpers
                                      (_save_solo_game, _clear_solo_game,
                                      _save_pvp_match, _clear_pvp_match,
                                      _save_game_state,
                                      _save_tournament_entry,
                                      _clear_tournament_entry,
                                      _is_solo_game,
                                      _pvp_canonical_user_id,
                                      _serialize_pvp_hand)
  views/blackjack/__init__.py       — re-exports
  views/blackjack/embeds.py         — _game_embed, _tourn_embed,
                                      _update_tourn_embed
  views/blackjack/solo_view.py      — BlackjackView (solo + PvP hand)
  views/blackjack/pvp_view.py       — _ChallengeView, _start_pvp,
                                      _resolve_pvp
  views/blackjack/tournament_views  — _TournRegistrationView,
                                      _TournBlackjackView,
                                      _start_tourn_round,
                                      _check_tourn_done

Import direction (no cycles):
  cogs/blackjack/_state.py       imports nothing from cogs.blackjack_cog
  cogs/blackjack/_persistence.py imports from cogs.blackjack._state
  views/blackjack/*              imports from cogs.blackjack._state +
                                 cogs.blackjack._persistence
  cogs/blackjack_cog.py          imports from cogs.blackjack._state +
                                 cogs.blackjack._persistence +
                                 views.blackjack

Behavior preserved end-to-end:

  * !blackjack, !blackjack @user [bet], !bjtournament, !bjstart,
    !bjstatus all work unchanged
  * Solo / PvP / tournament flows preserve message ownership, edit
    behavior, and timeout semantics
  * cog_load recovery (_recover_blackjack_solo, _recover_blackjack_pvp,
    _recover_blackjack_tournament) and on_guild_remove still refund
    tournament entries while clearing solo/PvP rows without refund
  * Module dicts (_active, _pvp, _tournaments) are still shared
    objects across the cog + views (single import sources them once;
    Python's import cache preserves identity)
  * Identity contract surfaces unchanged (no PersistentView in this
    subsystem)
  * No new abstractions introduced — every name in the new modules
    existed in the pre-extraction cog

Test patch path updates (mechanical, intent preserved):

  Tests previously patched ``cogs.blackjack_cog.game_state_service.save``
  and ``.clear`` to assert the persistence helpers' Postgres writes.
  After the move, the helpers live in cogs.blackjack._persistence and
  look up game_state_service in THAT module's namespace — so patches
  must target the new module path.

    cogs.blackjack_cog.game_state_service.save   → cogs.blackjack._persistence.game_state_service.save  (10 sites)
    cogs.blackjack_cog.game_state_service.clear  → cogs.blackjack._persistence.game_state_service.clear (3 sites)
    cogs.blackjack_cog._save_solo_game           → cogs.blackjack._persistence._save_solo_game           (1 site)
    cogs.blackjack_cog._save_pvp_match           → cogs.blackjack._persistence._save_pvp_match           (1 site)

  Updated via mechanical sed across three test files:
    tests/unit/cogs/test_blackjack_solo_persistence.py
    tests/unit/cogs/test_blackjack_pvp_persistence.py
    tests/unit/cogs/test_blackjack_tournament_persistence.py

  Test INTENT preserved exactly — same call assertions, same fixture
  shapes, same kwargs verification.  ``from cogs.blackjack_cog import _Game``
  etc. still works via the back-compat re-exports at the bottom of
  cogs/blackjack_cog.py.

What is NOT in this PR (intentionally deferred):

  * Engine extension (audit §3.4 proposed adding GameState /
    PvPState / TournamentRoundState dataclasses to
    services/blackjack_engine.py).  This stage relocates existing
    code only; engine extension is its own design exercise.
  * Cross-subsystem services (tournament_lock_service,
    tournament_registration_service from audit §6.5 / §6.2(a)).
    Those need both rps and blackjack consumers landing together —
    out of scope for this PR per the controlled-mode discipline
    that forbids transitional architecture half-complete across
    unrelated subsystems.

Validation:

  * pytest tests/unit/                            837 / 837 pass
  * test_blackjack_solo_persistence.py            9   / 9   pass
  * test_blackjack_pvp_persistence.py             8   / 8   pass
  * test_blackjack_tournament_persistence.py      8   / 8   pass
  * Identity contract STRICT mode                 green
  * INV-A..INV-N                                  green
  * black, isort, ruff, mypy (whole tree, 178 source files)  clean
  * Import smoke test: 35 symbols round-trip through re-exports

Phase A cog-size scoreboard after this PR:

  blackjack_cog.py             650  ✓ now under S4.6 ceiling
  rps_tournament_cog.py       1182  ⚠ further extraction recommended
  diagnostic_cog.py            544  ✓
  counting_cog.py              654  warn-tier
  role_cog.py                  620  warn-tier
  moderation_cog.py            212  ✓
  help_cog.py                  506  warn-tier
  chain_cog.py                 504  warn-tier
  xp_cog.py                    110  ✓

After this PR, only ``rps_tournament_cog.py`` is over 800 LOC.
S4.6 (soft size invariant CI gate) can land after one more rps
extraction pass or after the cross-subsystem services land.
